Crate pallet_governed_map

Pallet exposing the key-value pairs sourced from the Cardano ledger as part of the Governed Map feature.

Purpose of this pallet

This pallet stores the most recent state of the key-value pairs in the Governed Map on Cardano for use by other runtime components. It also exposes hooks for other components to be notified when particular key-value pair is inserted, updated or deleted.

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_governed_map