pallet_governed_map

Module pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

[Event] and Error, which are listed among the Enums.

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

Error type used by this pallet’s extrinsics

Constants

PALLET_VERSION

Current pallet version

Traits

Config

Configuration trait of this pallet.

Type Aliases

Changes

Governed Map change list

Initialized

Stores the initialization state of the pallet

LastUpdateBlock

Stores the block number of the last time mapping changes were registered

MainChainScripts

Cardano address of the Governed Map validator.

MapKey

Governed Map key type

MapValue

Governed Map value type

Mapping

Stores the latest state of the Governed Map that was observed on Cardano.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.