pallet_governed_map::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

register_changes

Inherent to register any changes in the state of the Governed Map on Cardano compared to the state currently stored in the pallet.

set_main_chain_scripts

Changes the address of the Governed Map validator used for observation.